Windows Alt Code Changes
So the Unicode Characters are not consistent but Windows All Codes are.

Developed an algorithm that allows that simulation of Num Lock + Left Alt + Keypad Entry for a bunch of A, E, I, O, U with acute accents and AE.

Also added in N with tilda.  Moved the massdrop control buttons to F1/F2 with Function pressed and added Escape as a standard key also with function pressed